scaffolding-node-modules: add reference to my collection
Referencing collection exposes it to anyone reading this repository.
Downside is updating structure of collection. Then I have to update my
reference to it here.

This referencing is required because collection of research is my main
point of work, everything is stored there.
But not all research can be done in stacked repository. Other work can
tamper with current one.